test(combat): end-to-end +penetration modifier proof; roadmap updates (#767)
