+++ docs/roadmap/seed-and-integration-tests.md
@@ -54,6 +54,7 @@ These can overlap at the boundaries — for example, the magic-cluster seed help
 | ✅ **1.8** — Cantrip starter catalog via `seed_cantrip_starter_catalog()`: 5 TechniqueStyle rows, 6 EffectType rows, 25 Cantrip rows (5 styles × 5 archetypes), 5 PROSPECT Path rows. | CG magic stage works on a fresh dev install. New characters can pick a starter cantrip. |
 | ✅ **1.9** — `seed_magic_dev()` master function in `integration_tests/game_content/magic.py` that calls 1.1, 1.2, 1.3, 1.8 plus `author_reference_corruption_content()` and `MagicContent.create_all()`. Two idempotency fixes: `_make_magical_endurance_check_type()` switched from factory to direct ORM (converges with seed_magic_config's row); `MagicContent.create_all()` switched to get_or_create (no duplicate Technique/ActionEnhancement rows on re-run). 7 new tests covering idempotency and convergence. | One call seeds the entire magic cluster. Ready to become the magic-cluster contribution to Phase 3's `seed_dev_database()`. |
 | ✅ **1.10** — Magic-story pipeline slice (resonance-environment-based): `AffinityInteraction` rows (9 directed pairs) + `ResonanceEnvironmentConfig` singleton seeded into `seed_magic_dev()`; rooms seeded via `tag_room_resonance` (cascade substrate). Universal-path redesign (2026-05-16): dropped "Magically Attuned" condition + universal trigger + universal flow seeding; replaced with OPPOSED `ConsequencePool`/`Consequence`/`ConsequenceEffect` rows for the celestial-place pairs (#4/#7) and ALIGNED `ResonanceAlignmentBoonTier` rows with named buff `ConditionTemplate`s for the abyssal/abyssal pair (#5). Pipeline test `test_magic_story_pipeline.py` — subtests cover OPPOSED backfire (4 `CheckOutcome` tiers × 2 place-intensity bands via the real `resonance_environment_for_cast` orchestrator call), ALIGNED presence buff apply/clear/swap via `refresh_resonance_alignment`, Quiescent (no `CharacterAura`) inert case, CASTER_DOMINANT stub (fill-in for deferred defilement), and second-earner Discovery. `setUp` applies no condition — exercises the production path directly. | Magic-story resonance-environment pipeline provably works seed-through-pipeline on the production code path. The universal-path redesign removes the marker-condition anti-pattern; magic-capability is now derived from `CharacterAura` presence. The primitive (`evaluate_resonance_environment`) is the shared substrate for deferred presence-escalation, defilement (CASTER_DOMINANT), and combat clash-of-wills. |
+| ✅ **1.11** — `game_content/combat.py` now exists with `seed_penetration_contest()` composed into `seed_magic_dev()` step 9 (#767). Seeds the trait-weighted penetration `CheckType` (willpower × 1.00 + intellect × 0.50), the 4-rung `PenetrationOutcomeFactor` ladder (bounce / partial / clean / overpenetration), and the check-scoped penetration `ModifierTarget` so caster-side `CharacterModifier` rows (e.g. "+penetration vs warded foes" distinctions) flow through `collect_check_modifiers` into the penetration contest's `extra_modifiers`. Generic `CHARACTER` source added to `collect_check_modifiers` — any check with a linked `ModifierTarget` gains `+X to <check>` support. Penetration check routed through the shared modifier seam in `_apply_penetration`. | Any check-scoped distinction (e.g. Wardbreaker) automatically applies to the penetration contest without additional wiring. |
 
 ### Phase 1 not in scope
