+++ src/world/combat/tests/test_penetration_seed.py
@@ -6,6 +6,7 @@ target_check_type), and idempotency of every wire function.
 """
 
 from decimal import Decimal
+from unittest.mock import MagicMock, patch
 
 from django.test import TestCase
 
@@ -15,7 +16,21 @@ from world.combat.factories import (
     wire_penetration_check_type,
     wire_penetration_modifier_target,
 )
+from world.combat.tests.penetration_helpers import _build_resolver, _ledger
+from world.conditions.factories import (
+    DamageSuccessLevelMultiplierFactory,
+    wire_penetration_factors,
+)
+from world.distinctions.factories import (
+    CharacterDistinctionFactory,
+    DistinctionEffectFactory,
+    DistinctionFactory,
+)
 from world.mechanics.constants import CHECK_CATEGORY_NAME
+from world.mechanics.factories import (
+    CharacterModifierFactory,
+    DistinctionModifierSourceFactory,
+)
 from world.mechanics.models import ModifierTarget
 
 
@@ -72,3 +87,45 @@ class WirePenetrationModifierTargetTests(TestCase):
         second = wire_penetration_modifier_target()
         self.assertEqual(first.pk, second.pk)
         self.assertEqual(ModifierTarget.objects.filter(name=PENETRATION_CHECK_TYPE_NAME).count(), 1)
+
+
+class PenetrationModifierContestTests(TestCase):
+    """A +penetration CharacterModifier feeds the contest end-to-end (#767)."""
+
+    @classmethod
+    def setUpTestData(cls) -> None:
+        DamageSuccessLevelMultiplierFactory(
+            min_success_level=2, multiplier=Decimal("1.00"), label="Modifier Full"
+        )
+        DamageSuccessLevelMultiplierFactory(
+            min_success_level=1, multiplier=Decimal("0.50"), label="Modifier Partial"
+        )
+        wire_penetration_factors()
+        wire_penetration_check_type()
+
+    def setUp(self) -> None:
+        from evennia.utils.idmapper import models as idmapper_models
+
+        idmapper_models.flush_cache()
+
+    def test_character_modifier_raises_penetration_extra_modifiers(self) -> None:
+        target = wire_penetration_modifier_target()
+        resolver = _build_resolver(barrier_strength=10)
+        distinction = DistinctionFactory(name="Wardbreaker")
+        effect = DistinctionEffectFactory(distinction=distinction, target=target, value_per_rank=3)
+        source = DistinctionModifierSourceFactory(
+            distinction_effect=effect,
+            character_distinction=CharacterDistinctionFactory(distinction=distinction),
+        )
+        CharacterModifierFactory(
+            character=resolver.participant.character_sheet,
+            value=3,
+            source=source,
+            target=target,
+        )
+
+        with patch("world.combat.services.perform_check") as mock_pen:
+            mock_pen.return_value = MagicMock(success_level=1)
+            resolver(power=20, ledger=_ledger(20))
+
+        self.assertEqual(mock_pen.call_args.kwargs["extra_modifiers"], 3)
